backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-multipart==0.0.9

# ── Database ───────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ── Cache / Task Queue ─────────────────────────────────────────
redis==5.0.4
celery[redis]==5.3.6

# ── Auth ───────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Config ─────────────────────────────────────────────────────
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── Network Discovery ──────────────────────────────────────────
python-nmap==0.7.1
scapy==2.5.0
pysnmp==6.1.2
netaddr==1.3.0
mac-vendor-lookup==0.1.12
zeroconf==0.132.2          # mDNS/Bonjour discovery

# ── Protocol Probes ────────────────────────────────────────────
impacket==0.12.0            # SMB/NetBIOS probe

# ── AI Agent ───────────────────────────────────────────────────
openai==1.35.0              # Ollama OpenAI-compatible API + optional OpenAI
anthropic==0.28.0           # Optional: Anthropic API fallback

# ── Utilities ──────────────────────────────────────────────────
httpx[http2]==0.27.0        # Async HTTP for probes (HTTP/2 support)
aiofiles==23.2.1
structlog==24.1.0
